json_exists_query: add logs
Log this as warn because quickwit is configured to print logs with
  tantivy=WARN

Log this with self.full_path so that each full_path will have a unique
rate-limit (and not a global rate-limit for all exists queries).

Log with limit_per_sec=1 granularity so that hopefully every search will lead to
some logs.